+++ messaging/meowsenger/src/main/java/meow/alxnko/meowsenger/repository/ChatRepository.java
@@ -9,12 +9,12 @@ import org.springframework.stereotype.Repository;
 import meow.alxnko.meowsenger.model.Chat;
 
 @Repository
-public interface ChatRepository extends JpaRepository<Chat, Integer> { // Changed from Long to Integer
+public interface ChatRepository extends JpaRepository<Chat, Long> {
     
     @Query("SELECT c FROM Chat c WHERE :userId MEMBER OF c.participantIds")
-    List<Chat> findByParticipantId(Integer userId); // Changed from Long to Integer
+    List<Chat> findByParticipantId(Long userId);
     
     // Check if a user is part of a chat
     @Query("SELECT CASE WHEN COUNT(c) > 0 THEN true ELSE false END FROM Chat c WHERE c.id = :chatId AND :userId MEMBER OF c.participantIds")
-    boolean isUserInChat(Integer chatId, Integer userId); // Changed from Long to Integer
+    boolean isUserInChat(Long chatId, Long userId);
 }
